Decided when to activate the magnetic rail
I wrote a script that notices when the drone is close to the rover and then
actuates... something. I have no idea how to use odroid, and I'm pretty sure
all of the numbers are wrong since I don't know how to test them. Theoretically
there is a distance sensor on pin 0 and the actuator for the magnetic rail on
pin 1. I really don't know anything about odroids.